+++ eslint/eslint.config.js
@@ -1,0 +1,3 @@
+const nextEslintConfig = require('./eslint.next.config');
+
+module.exports = [...nextEslintConfig];