+++ packages/compiler/src/compilation/setup-compilation.unit.test.ts
@@ -108,6 +108,7 @@ describe('setupCompilation', () => {
       },
       host: expect.any(Object),
       rootNames: ['main.ts'],
+      componentStylesheetBundler: expect.any(Object),
     });
   });