+++ packages/compiler/src/compilation/setup-with-paralell-compilation.ts
@@ -11,11 +11,12 @@ export async function setupCompilationWithParallelCompilation(
   config: Pick<RsbuildConfig, 'source'>,
   options: SetupCompilationOptions
 ) {
-  const { rootNames, compilerOptions } = await setupCompilation(
-    config,
-    options
+  const { rootNames, compilerOptions, componentStylesheetBundler } =
+    await setupCompilation(config, options);
+  const parallelCompilation = new ParallelCompilation(
+    options.jit ?? false,
+    options.hasServer === false
   );
-  const parallelCompilation = new ParallelCompilation(options.jit ?? false);
   const fileReplacements: Record<string, string> =
     options.fileReplacements.reduce((r, f) => {
       r[f.replace] = f.with;
@@ -29,8 +30,7 @@ export async function setupCompilationWithParallelCompilation(
         ...compilerOptions,
         fileReplacements,
         modifiedFiles: new Set(rootNames),
-        transformStylesheet: (styles) =>
-          Promise.resolve(styleTransform(styles)),
+        transformStylesheet: styleTransform(componentStylesheetBundler),
         processWebWorker(workerFile: string) {
           return transformFileSync(workerFile).code;
         },